Reorganise files to separate layout, callbacks and utils.
